add theme service
